Initial migration to library; add decryption operation
